Remove unused code and moved baker into Editor
