+++ Packages/jp.keijiro.bodypix/Editor/Phase1ModelBuilder.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 82f1085064b343f6af5288d27219252a
+guid: 549f5892a5be4609a861c850a91c58c0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
