+++ Packages/jp.keijiro.bodypix/Runtime/ResourceSet.cs.meta
@@ -5,8 +5,6 @@ MonoImporter:
   serializedVersion: 2
   defaultReferences:
   - model: {fileID: 5022602860645237092, guid: ff475820fb5ea4db1b073da3f6892ddb, type: 3}
-  - preprocess: {fileID: 7200000, guid: 065b9d2f5e685444881e88945bf0492b, type: 3}
-  - mask: {fileID: 7200000, guid: c22c946b1ee1744b6b57e01fbedd3fd4, type: 3}
   - keypoints: {fileID: 7200000, guid: 7f8a40b27a6624527831abb31e5be62b, type: 3}
   executionOrder: 0
   icon: {instanceID: 0}
