--- Packages/jp.keijiro.bodypix/Shaders/Mask.compute.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c22c946b1ee1744b6b57e01fbedd3fd4
-ComputeShaderImporter:
-  externalObjects: {}
-  preprocessorOverride: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
